Fix test fixtures for test_convert
