+++ tests/convert/test_area.py
@@ -38,35 +38,6 @@ def regions():
 
 
 @fixture(scope='function')
-def regions_half_squares():
-    """Return two adjacent square regions::
-
-        |```|```|
-        | A | B |
-        |...|...|
-
-    """
-    return RegionSet('half_squares', [
-        {
-            'type': 'Feature',
-            'properties': {'name': 'a'},
-            'geometry': {
-                'type': 'Polygon',
-                'coordinates': [[[0, 0], [0, 1], [1, 1], [1, 0]]]
-            }
-        },
-        {
-            'type': 'Feature',
-            'properties': {'name': 'b'},
-            'geometry': {
-                'type': 'Polygon',
-                'coordinates': [[[0, 1], [0, 2], [1, 2], [1, 1]]]
-            }
-        },
-    ])
-
-
-@fixture(scope='function')
 def regions_single_half_square():
     """Return single half-size square region::
 
@@ -82,27 +53,6 @@ def regions_single_half_square():
             'geometry': {
                 'type': 'Polygon',
                 'coordinates': [[[0, 0], [0, 1], [1, 1], [1, 0]]]
-            }
-        }
-    ])
-
-
-@fixture(scope='function')
-def regions_rect():
-    """Return single region covering 2x1 area::
-
-        |```````|
-        |   0   |
-        |.......|
-
-    """
-    return RegionSet('rect', [
-        {
-            'type': 'Feature',
-            'properties': {'name': 'zero'},
-            'geometry': {
-                'type': 'Polygon',
-                'coordinates': [[[0, 0], [0, 2], [1, 2], [1, 0]]]
             }
         }
     ])
