+++ tests/convert/test_convert.py
@@ -1,10 +1,29 @@
 import numpy as np
+from pytest import fixture
 from smif.convert import SpaceTimeUnitConvertor
+from smif.convert.interval import IntervalSet
+
+
+@fixture(scope='function')
+def convertor(regions_rect, regions_half_squares, months, seasons, twenty_four_hours, one_day,
+              remap_months):
+    """Convertor with dimensions pre-registered for use
+    """
+    convertor = SpaceTimeUnitConvertor()
+    convertor.regions.register(regions_rect)
+    convertor.regions.register(regions_half_squares)
+    convertor.intervals.register(IntervalSet('months', months))
+    convertor.intervals.register(IntervalSet('seasons', seasons))
+    convertor.intervals.register(IntervalSet('hourly_day', twenty_four_hours))
+    convertor.intervals.register(IntervalSet('one_day', one_day))
+    convertor.intervals.register(IntervalSet('remap_months', remap_months))
+    return convertor
 
 
 class TestSpaceTimeUnitConvertor_TimeOnly:
-
-    def test_one_region_pass_through_time(self):
+    """No change in regions
+    """
+    def test_one_region_pass_through_time(self, convertor):
         """Only one region, 12 months, neither space nor time conversion is required
         """
         data = np.array([[
@@ -21,7 +40,6 @@ class TestSpaceTimeUnitConvertor_TimeOnly:
             30,
             31
         ]], dtype=float)
-        convertor = SpaceTimeUnitConvertor()
         actual = convertor.convert(
             data,
             'half_squares',
@@ -33,7 +51,7 @@ class TestSpaceTimeUnitConvertor_TimeOnly:
         )
         assert np.allclose(actual, data)
 
-    def test_one_region_time_aggregation(self):
+    def test_one_region_time_aggregation(self, convertor):
         """Only one region, time aggregation is required
         """
         data = np.array([[
@@ -58,8 +76,6 @@ class TestSpaceTimeUnitConvertor_TimeOnly:
             30 + 31 + 30
         ]], dtype=float)  # area a, seasons 1-4
 
-        convertor = SpaceTimeUnitConvertor()
-
         actual = convertor.convert(
             data,
             'half_squares',
@@ -71,7 +87,7 @@ class TestSpaceTimeUnitConvertor_TimeOnly:
         )
         assert np.allclose(actual, expected)
 
-    def test_two_region_time_aggregation(self):
+    def test_two_region_time_aggregation(self, convertor):
         """Two regions, time aggregation by region is required
         """
         data = np.array([
@@ -107,7 +123,6 @@ class TestSpaceTimeUnitConvertor_TimeOnly:
             ]
         ], dtype=float)
 
-        convertor = SpaceTimeUnitConvertor()
         actual = convertor.convert(
             data,
             'half_squares',
@@ -135,12 +150,10 @@ class TestSpaceTimeUnitConvertor_TimeOnly:
 
         assert np.allclose(actual, expected)
 
-    def test_one_region_convert_from_hour_to_day(self):
+    def test_one_region_convert_from_hour_to_day(self, convertor):
         """One region, time aggregation required
         """
-
         data = np.ones((1, 24))  # area a, hours 0-23
-        convertor = SpaceTimeUnitConvertor()
         actual = convertor.convert(
             data,
             'half_squares',
@@ -155,8 +168,9 @@ class TestSpaceTimeUnitConvertor_TimeOnly:
 
 
 class TestRemapConversion:
-
-    def test_remap_timeslices_to_months(self):
+    """Repeating values
+    """
+    def test_remap_timeslices_to_months(self, convertor):
         """One region, time remapping required
         """
         data = np.array([[
@@ -165,8 +179,6 @@ class TestRemapConversion:
             1,  # summer month
             1  # autumn month
         ]], dtype=float)
-
-        convertor = SpaceTimeUnitConvertor()
         actual = convertor.convert(
             data,
             'half_squares',
@@ -181,12 +193,10 @@ class TestRemapConversion:
                               0.98901099, 1.02197802, 0.98901099, 1.03333333]])
         assert np.allclose(actual, expected)
 
-    def test_remap_months_to_timeslices(self, monthly_data, remap_month_data):
+    def test_remap_months_to_timeslices(self, convertor, monthly_data, remap_month_data):
         """One region, time remapping required
         """
         data = monthly_data
-
-        convertor = SpaceTimeUnitConvertor()
         actual = convertor.convert(
             data,
             'half_squares',
@@ -201,15 +211,12 @@ class TestRemapConversion:
 
 
 class TestSpaceTimeUnitConvertor_RegionOnly:
-
-    def test_half_to_one_region_pass_through_time(self):
+    """No change in intervals
+    """
+    def test_half_to_one_region_pass_through_time(self, convertor):
         """Convert from half a region to one region, pass through time
-
         """
-
         data = np.ones((2, 12)) / 2  # area a,b, months 1-12
-
-        convertor = SpaceTimeUnitConvertor()
         actual = convertor.convert(
             data,
             'half_squares',
@@ -222,13 +229,11 @@ class TestSpaceTimeUnitConvertor_RegionOnly:
         expected = np.ones((1, 12))  # area zero, months 1-12
         assert np.allclose(actual, expected)
 
-    def test_one_region_convert_from_hour_to_day(self):
+    def test_one_region_convert_from_hour_to_day(self, convertor):
         """Two regions aggregated to one, one interval
         """
 
         data = np.array([[24], [24]])  # area a,b, single interval
-
-        convertor = SpaceTimeUnitConvertor()
         actual = convertor.convert(
             data,
             'half_squares',
@@ -243,11 +248,10 @@ class TestSpaceTimeUnitConvertor_RegionOnly:
 
 
 class TestSpaceTimeUnitConvertorBoth:
-
-    def test_convert_both_space_and_time(self):
+    """Change both regions and intervals
+    """
+    def test_convert_both_space_and_time(self, convertor):
         data = np.ones((2, 12)) / 2  # area a,b, months 1-12
-
-        convertor = SpaceTimeUnitConvertor()
         actual = convertor.convert(
             data,
             'half_squares',
@@ -257,21 +261,5 @@ class TestSpaceTimeUnitConvertorBoth:
             'm',
             'm'
         )
-        expected = np.ones((1, 4)) * 3  # area zero, seasons 1-4
-        assert np.allclose(actual, expected)
-
-
-class TestConvertor:
-
-    def test_convertor(self):
-        data = np.ones((2, 12)) / 2  # area a,b, months 1-12
-        convertor = SpaceTimeUnitConvertor()
-        actual = convertor.convert(data,
-                                   'half_squares',
-                                   'rect',
-                                   'months',
-                                   'seasons',
-                                   'm',
-                                   'm')
         expected = np.ones((1, 4)) * 3  # area zero, seasons 1-4
         assert np.allclose(actual, expected)
